Define another cylindrical detector from Rigaku
